# autotools: add support for 'unity' builds, enable in CI