--- a/.github/workflows/awslc.yml
+++ b/.github/workflows/awslc.yml
@@ -37,7 +37,6 @@ concurrency:
 permissions: {}
 
 env:
-  CURL_CI: github
   MAKEFLAGS: -j 5
 
   # renovate: datasource=github-tags depName=awslabs/aws-lc versioning=semver registryUrl=https://github.com