--- a/.github/workflows/http3-linux.yml
+++ b/.github/workflows/http3-linux.yml
@@ -41,6 +41,7 @@ concurrency:
 permissions: {}
 
 env:
+  CURL_CI: github
   MAKEFLAGS: -j 5
   # handled in renovate.json
   openssl3-version: openssl-3.3.0
@@ -472,4 +473,3 @@ jobs:
         name: 'run pytest'
         env:
           TFLAGS: "${{ matrix.build.tflags }}"
-          CURL_CI: github