--- a/.github/workflows/http3-linux.yml
+++ b/.github/workflows/http3-linux.yml
@@ -41,7 +41,6 @@ concurrency:
 permissions: {}
 
 env:
-  CURL_CI: github
   MAKEFLAGS: -j 5
   # handled in renovate.json
   openssl3-version: openssl-3.3.0
@@ -452,7 +451,7 @@ jobs:
       - run: autoreconf -fi
         name: 'autoreconf'
 
-      - run: ./configure --disable-dependency-tracking ${{ matrix.build.configure }}
+      - run: ./configure --disable-dependency-tracking --enable-unity ${{ matrix.build.configure }}
         name: 'configure'
 
       - run: make V=1
@@ -473,3 +472,4 @@ jobs:
         name: 'run pytest'
         env:
           TFLAGS: "${{ matrix.build.tflags }}"
+          CURL_CI: github