--- a/.github/workflows/linux32.yml
+++ b/.github/workflows/linux32.yml
@@ -40,6 +40,7 @@ concurrency:
 permissions: {}
 
 env:
+  CURL_CI: github
   MAKEFLAGS: -j 5
 
 jobs: