--- a/.github/workflows/linux32.yml
+++ b/.github/workflows/linux32.yml
@@ -40,7 +40,6 @@ concurrency:
 permissions: {}
 
 env:
-  CURL_CI: github
   MAKEFLAGS: -j 5
 
 jobs:
@@ -70,7 +69,9 @@ jobs:
       - run: autoreconf -fi
         name: 'autoreconf'
 
-      - run: ./configure --disable-dependency-tracking --enable-warnings --enable-werror ${{ matrix.build.configure }}
+      - run: |
+          ./configure --disable-dependency-tracking --enable-unity --enable-warnings --enable-werror \
+            ${{ matrix.build.configure }}
         name: 'configure'
 
       - run: make V=1