--- a/.github/workflows/non-native.yml
+++ b/.github/workflows/non-native.yml
@@ -35,6 +35,9 @@ concurrency:
 
 permissions: {}
 
+env:
+  CURL_CI: github
+
 jobs:
   netbsd:
     name: 'NetBSD (cmake, openssl, clang)'