--- a/.github/workflows/non-native.yml
+++ b/.github/workflows/non-native.yml
@@ -35,9 +35,6 @@ concurrency:
 
 permissions: {}
 
-env:
-  CURL_CI: github
-
 jobs:
   netbsd:
     name: 'NetBSD (cmake, openssl, clang)'
@@ -135,7 +132,7 @@ jobs:
               pkgconf brotli openldap26-client libidn2 libnghttp2 nghttp2 stunnel py311-impacket
             autoreconf -fi
             export CC='${{ matrix.compiler }}'
-            mkdir bld && cd bld && ../configure --enable-debug --enable-warnings --enable-werror \
+            mkdir bld && cd bld && ../configure --enable-unity --enable-debug --enable-warnings --enable-werror \
               --prefix="${HOME}"/install \
               --enable-websockets \
               --with-openssl \
@@ -193,7 +190,7 @@ jobs:
           run: |
             ln -s /usr/bin/gcpp /usr/bin/cpp  # Some tests expect `cpp`, which is named `gcpp` in this env.
             autoreconf -fi
-            mkdir bld && cd bld && ../configure --enable-debug --enable-warnings --enable-werror \
+            mkdir bld && cd bld && ../configure --enable-unity --enable-debug --enable-warnings --enable-werror \
               --prefix="${HOME}"/install \
               --enable-websockets \
               --with-openssl \