--- a/.github/workflows/torture.yml
+++ b/.github/workflows/torture.yml
@@ -41,6 +41,7 @@ concurrency:
 permissions: {}
 
 env:
+  CURL_CI: github
   MAKEFLAGS: -j 5
 
 jobs: