--- a/.github/workflows/torture.yml
+++ b/.github/workflows/torture.yml
@@ -41,7 +41,6 @@ concurrency:
 permissions: {}
 
 env:
-  CURL_CI: github
   MAKEFLAGS: -j 5
 
 jobs:
@@ -75,7 +74,9 @@ jobs:
       - run: autoreconf -fi
         name: 'autoreconf'
 
-      - run: ./configure --disable-dependency-tracking --enable-warnings --enable-werror ${{ matrix.build.configure }}
+      - run: |
+          ./configure --disable-dependency-tracking --enable-unity --enable-warnings --enable-werror \
+            ${{ matrix.build.configure }}
         name: 'configure'
 
       - run: make V=1