--- a/lib/Makefile.am
+++ b/lib/Makefile.am
@@ -71,7 +71,9 @@ AM_CFLAGS =
 include Makefile.inc
 
 if USE_UNITY
-curl_EXCLUDE =
+# Keep these separate to avoid duplicate definitions when linking libtests
+# in static mode.
+curl_EXCLUDE = curl_threads.c timediff.c warnless.c
 if DEBUGBUILD
 # We must compile these sources separately to avoid memdebug.h redefinitions
 # applying to them.