--- a/scripts/Makefile.am
+++ b/scripts/Makefile.am
@@ -23,7 +23,7 @@
 ###########################################################################
 
 EXTRA_DIST = coverage.sh completion.pl firefox-db2pem.sh checksrc.pl    \
- mk-ca-bundle.pl schemetable.c cd2nroff nroff2cd cdall cd2cd managen \
+ mk-ca-bundle.pl mk-unity.pl schemetable.c cd2nroff nroff2cd cdall cd2cd managen \
  dmaketgz maketgz release-tools.sh verify-release cmakelint.sh
 
 ZSH_FUNCTIONS_DIR = @ZSH_FUNCTIONS_DIR@