@vszakats 2024-09-20
This new "unity" option for autotools is disabled by default, except if the env `CURL_CI` is
set, which enables it by default. This env already had some use in GHA, this PR sets it for
most workflows globally. In a few workflows it's enabled with `--enable-unity` and in two
workflows: distro and AWC-LC it's intentionally not enabled.

It does work fine in distro, but it's probably best to use the default there.

AWC-LC is a strange case where the build is already faster than elsewhere for some reason,
and unity makes the build a few seconds slower. (Perhaps something obvious, but I don't
know why this is, and I could not see similar result in other jobs.)

This PR is ready for merge.