@vszakats 2024-09-20
Well, the `CURL_CI` trick didn't save much editing, while it reduced
transparency and local reproducibility, so went back to just enabling
`--enable-unity` manually where needed. There is no new "special
case" now, `--enable-unity` is just disabled by default, always.